requirements-pip.txt
torch==2.10.0
xarray==2026.1.0
PyYAML==6.0.3